+++ frontend/wailsjs/go/main/App.d.ts
@@ -4,11 +4,17 @@ import {main} from '../models';
 
 export function GetAnnouncements():Promise<Array<main.AnnouncementSettings>>;
 
+export function GetKnowledge():Promise<main.KnowledgeSettings>;
+
 export function GetLogs():Promise<Array<string>>;
 
 export function GetSettings():Promise<main.ControlSettings>;
 
+export function ResetKnowledgeTemplate():Promise<main.KnowledgeSettings>;
+
 export function SaveAnnouncements(arg1:Array<main.AnnouncementSettings>):Promise<void>;
+
+export function SaveKnowledge(arg1:main.KnowledgeSettings):Promise<void>;
 
 export function SaveSettings(arg1:main.ControlSettings):Promise<void>;