+++ frontend/wailsjs/go/main/App.js
@@ -6,6 +6,10 @@ export function GetAnnouncements() {
   return window['go']['main']['App']['GetAnnouncements']();
 }
 
+export function GetKnowledge() {
+  return window['go']['main']['App']['GetKnowledge']();
+}
+
 export function GetLogs() {
   return window['go']['main']['App']['GetLogs']();
 }
@@ -14,8 +18,16 @@ export function GetSettings() {
   return window['go']['main']['App']['GetSettings']();
 }
 
+export function ResetKnowledgeTemplate() {
+  return window['go']['main']['App']['ResetKnowledgeTemplate']();
+}
+
 export function SaveAnnouncements(arg1) {
   return window['go']['main']['App']['SaveAnnouncements'](arg1);
+}
+
+export function SaveKnowledge(arg1) {
+  return window['go']['main']['App']['SaveKnowledge'](arg1);
 }
 
 export function SaveSettings(arg1) {